% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquarium-package.R
\docType{package}
\name{aquarium}
\alias{aquarium}
\alias{aquarium-package}
\title{Validate 'Phishing' 'URLs' with the 'PhishTank' Service}
\description{
'PhishTank' <www.phishtank.com> is a free community site where anyone can submit,
verify, track and share 'phishing' data. Methods are provided to test if a 'URL' is classsified
as a 'phishing' site and to download aggregated 'phishing' 'URL' databases.
}
\author{
Bob Rudis (bob@rud.is)
}
