% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-url.R
\name{pt_check_url}
\alias{pt_check_url}
\title{Check an individual URL against PhishTank's database}
\usage{
pt_check_url(url, phishtank_api_key = Sys.getenv("PHISHTANK_API_KEY"))
}
\arguments{
\item{url}{URL to test}

\item{phishtank_api_key}{a PhishTank \href{http://www.phishtank.com/api_register.php}{API key}}
}
\value{
data frame (tibble)
}
\description{
Check an individual URL against PhishTank's database
}
\note{
You \emph{need} an \href{http://www.phishtank.com/api_register.php}{API key}. Anonymous usage is not
supported by this package.
}
\references{
\url{https://www.phishtank.com/api_info.php}
}
