% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-db.R
\name{pt_read_db}
\alias{pt_read_db}
\title{Retrieve a complete copy of the current PhishTank database}
\usage{
pt_read_db(phishtank_api_key = Sys.getenv("PHISHTANK_API_KEY"),
  .progress = TRUE)
}
\arguments{
\item{phishtank_api_key}{a PhishTank \href{http://www.phishtank.com/api_register.php}{API key}}

\item{.progress}{if \code{TRUE}, display a download progress bar}
}
\description{
Retrieve a complete copy of the current PhishTank database
}
\note{
You \emph{need} an \href{http://www.phishtank.com/api_register.php}{API key}. Anonymous usage is not
supported by this package.
}
\references{
\url{https://www.phishtank.com/developer_info.php}
}
